feat: :sparkles: Se agrega lógica para obtener un producto y eliminar un producto

+++ apps/teslo-shop/src/app/feature/products/products.service.ts
@@ -2,6 +2,7 @@ import {
   Injectable,
   InternalServerErrorException,
   Logger,
+  NotFoundException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
@@ -32,8 +33,11 @@ export class ProductsService {
   }
 
   async findOne(id: string) {
-    const product = await this.productRepository.findOne({ where: { id: id } });
-    return product;
+    const product = await this.getProduct(id);
+    return (
+      product ??
+      new NotFoundException(`The product with id ${id} was not found`)
+    );
   }
 
   update(id: number, updateProductDto: UpdateProductDto) {
@@ -41,12 +45,22 @@ export class ProductsService {
   }
 
   async remove(id: string) {
-    const product = await this.productRepository.delete(id);
-    return product;
+    const product = await this.getProduct(id);
+    if (product) {
+      await this.productRepository.delete(id);
+    } else {
+      throw new NotFoundException(`The product with id ${id} was not found`);
+    }
+    return `The product with id ${id} was successfully removed`;
   }
 
   private handleDBExceptions(error: any) {
     this.logger.error(error);
     throw new InternalServerErrorException('Ayuda!');
+  }
+
+  private async getProduct(id: string): Promise<Product | null> {
+    const product = await this.productRepository.findOne({ where: { id: id } });
+    return product;
   }
 }
